Реализация удаления точки маршрута и оптимизация сохранение, сортировки и папки utils
