+++ src/model/event-model.js
@@ -14,7 +14,7 @@ export default class EventsModel extends Observable {
   }
 
   updateEvent(updateType, update) {
-    const index = this.#events.findIndex((task) => task.id === update.id);
+    const index = this.#events.findIndex((event) => event.id === update.id);
 
     if (index === -1) {
       throw new Error('Can\'t update unexisting event');
@@ -39,10 +39,10 @@ export default class EventsModel extends Observable {
   }
 
   deleteEvent(updateType, update) {
-    const index = this.#events.findIndex((task) => task.id === update.id);
+    const index = this.#events.findIndex((event) => event.id === update.id);
 
     if (index === -1) {
-      throw new Error('Can\'t delete unexisting task');
+      throw new Error('Can\'t delete unexisting event');
     }
 
     this.#events = [
