+++ src/presenter/event-presenter.js
@@ -1,7 +1,8 @@
 import EventItemView from '../view/event-view.js';
 import FormEditView from '../view/form-edit-view.js';
 import { render, replace, remove } from '../framework/render.js';
-import { isEscapeKey, isEnterKey } from '../utils/utils.js';
+import { isEscapeKey } from '../utils/utils.js';
+import { isEventInPresent, isEventInPast, isEventInFuture } from '../utils/filter.js'
 import { Mode, UserAction, UpdateType } from '../const.js';
 
 export default class EventPresenter {
@@ -46,6 +47,7 @@ export default class EventPresenter {
       dataDestinations: this.#dataDestinations,
       onFormEditClick: this.#handleFormEditClick,
       onFormSubmit: this.#handleFormSubmit,
+      onDeleteClick: this.#handleDeleteClick,
     });
 
     if (prevEventComponent === null || prevFormEditComponent === null) {
@@ -94,14 +96,13 @@ export default class EventPresenter {
     if (isEscapeKey(evt)) {
       evt.preventDefault();
       this.#formEditComponent.reset(this.#event);
-      // console.log(1, this.#formEditComponent);
       this.#replaceFormToEvent();
     }
   };
 
   #handleFavoriteClick = () => {
     this.#handleDataChange(
-      UserAction.UPDATE_TASK,
+      UserAction.UPDATE_EVENT,
       UpdateType.MINOR,
       {...this.#event, isFavorite: !this.#event.isFavorite},
     );
@@ -109,23 +110,37 @@ export default class EventPresenter {
 
   #handleEditClick = () => {
     this.#replaceEventToForm();
+    document.removeEventListener('keydown', this.#escKeyDownHandler);
   };
 
   #handleFormEditClick = () => {
-    // this.#formEditComponent.reset(this.#event);
-    // console.log(2, this.#formEditComponent);
     this.#replaceFormToEvent();
+    document.addEventListener('keydown', this.#escKeyDownHandler);
   };
 
-  #handleFormSubmit = (event) => {
+  #handleFormSubmit = (update) => {
+    // Проверяем, поменялись ли в задаче данные, которые попадают под фильтрацию,
+    // а значит требуют перерисовки списка - если таких нет, это PATCH-обновление
+    const isMinorUpdate =
+      isEventInPast(this.#event.dateTo) !== isEventInPast(update.dateTo) ||
+      isEventInPresent(this.#event.dateFrom, this.#event.dateTo) !== isEventInPresent(update.dateFrom, update.dateTo) ||
+      isEventInFuture(this.#event.dateFrom) !== isEventInFuture(update.dateFrom);
+
     this.#handleDataChange(
-      UserAction.UPDATE_TASK,
-      UpdateType.MINOR,
-      event
+      UserAction.UPDATE_EVENT,
+      isMinorUpdate ? UpdateType.MINOR : UpdateType.PATCH,
+      update
     );
-    // this.#formEditComponent.reset(event);
-    // this.#eventComponent.reset(event);
-    // console.log(3, this.#formEditComponent);
+
     this.#replaceFormToEvent();
+    document.removeEventListener('keydown', this.#escKeyDownHandler);
+  };
+
+  #handleDeleteClick = (event) => {
+    this.#handleDataChange(
+      UserAction.DELETE_EVENT,
+      UpdateType.MINOR,
+      event,
+    );
   };
 }
